pyproject.toml
[project]
name = "rlm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "python-dotenv>=1.2.2",
    "rlms",
    "typer",
    "uvicorn",
]

[tool.uv]
package = false 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Submodules
[tool.uv.workspace]
members = ["rlm"]

[tool.uv.sources]
rlms = { workspace = true }
